BSAS-22 [+] modify 'int id' -> 'Long id'

+++ engine/src/main/java/com/jsoftware/platform/controller/MonthlyPickController.java
@@ -26,7 +26,7 @@ public class MonthlyPickController {
     }
 
     @GetMapping("/api/monthlyPick/{id}")
-    public MonthlyPick getMonthlyPick(@PathVariable("id") int id) {
+    public MonthlyPick getMonthlyPick(@PathVariable("id") Long id) {
         System.out.println(id);
         return Service.readMonthlyPick(id);
     }
